bot/requirements.txt
flask
requests
python-binance
openai>=1.0.0
python-dotenv
